refactor: rename node.py to abstract_node.py
