+++ docs/source/tutorials/04_peaking_duck.rst
@@ -102,7 +102,7 @@ Edit the following **five files** as described below:
    
          from typing import Any, Dict
          from datetime import datetime
-         from peekingduck.pipeline.nodes.node import AbstractNode
+         from peekingduck.pipeline.nodes.abstract_node import AbstractNode
          import sqlite3
    
          DB_FILE = "wave.db"           # name of database file
@@ -531,7 +531,7 @@ The ``filter_bbox.yml`` and ``filter_bbox.py`` files are shown below:
 
          from typing import Any, Dict
          import numpy as np
-         from peekingduck.pipeline.nodes.node import AbstractNode
+         from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
          class Node(AbstractNode):
