+++ peekingduck/pipeline/nodes/abstract_node.py
