+++ peekingduck/pipeline/nodes/augment/contrast.py
@@ -22,7 +22,7 @@ from typing import Any, Dict
 import cv2
 
 from peekingduck.pipeline.nodes.base import ThresholdCheckerMixin
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(ThresholdCheckerMixin, AbstractNode):
