+++ peekingduck/pipeline/nodes/dabble/bbox_to_btm_midpoint.py
@@ -16,7 +16,7 @@
 
 from typing import Any, Dict, Tuple
 
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(AbstractNode):
