+++ peekingduck/pipeline/nodes/dabble/check_large_groups.py
@@ -19,7 +19,7 @@ Checks if number of objects in a group exceeds a threshold.
 from collections import Counter
 from typing import Any, Dict, List
 
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(AbstractNode):
