+++ peekingduck/pipeline/nodes/dabble/check_nearby_objs.py
@@ -20,7 +20,7 @@ from typing import Any, Dict
 
 import numpy as np
 
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(AbstractNode):
