+++ peekingduck/pipeline/nodes/dabble/keypoints_to_3d_loc.py
@@ -20,7 +20,7 @@ from typing import Any, Dict
 
 import numpy as np
 
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 NOSE = 0
 LEFT_SHOULDER = 5
