+++ peekingduck/pipeline/nodes/draw/bbox.py
@@ -19,7 +19,7 @@ Draws bounding boxes over detected objects.
 from typing import Any, Dict
 
 from peekingduck.pipeline.nodes.draw.utils.bbox import draw_bboxes
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(AbstractNode):
