+++ peekingduck/pipeline/nodes/draw/blur_bbox.py
@@ -21,7 +21,7 @@ from typing import Any, Dict, List
 import cv2
 import numpy as np
 
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(AbstractNode):  # pylint: disable=too-few-public-methods
