+++ peekingduck/pipeline/nodes/draw/group_bbox_and_tag.py
@@ -22,7 +22,7 @@ import numpy as np
 
 from peekingduck.pipeline.nodes.draw.utils.bbox import draw_bboxes, draw_tags
 from peekingduck.pipeline.nodes.draw.utils.constants import TOMATO
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(AbstractNode):
