+++ peekingduck/pipeline/nodes/draw/zones.py
@@ -19,7 +19,7 @@ Draws the 2D boundaries of a zone.
 from typing import Any, Dict
 
 from peekingduck.pipeline.nodes.draw.utils.zone import draw_zones
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(AbstractNode):
