+++ peekingduck/pipeline/nodes/model/yolo.py
@@ -20,7 +20,7 @@ import cv2
 import numpy as np
 
 from peekingduck.pipeline.nodes.model.yolov4 import yolo_model
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(AbstractNode):
