+++ peekingduck/pipeline/nodes/model/yolox.py
@@ -19,7 +19,7 @@ from typing import Any, Dict
 import numpy as np
 
 from peekingduck.pipeline.nodes.model.yoloxv1 import yolox_model
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(AbstractNode):  # pylint: disable=too-few-public-methods
