+++ peekingduck/pipeline/nodes/node_template.py
@@ -18,7 +18,7 @@ Node template for creating custom nodes.
 
 from typing import Any, Dict
 
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(AbstractNode):
