+++ peekingduck/pipeline/nodes/output/screen.py
@@ -20,7 +20,7 @@ from typing import Any, Dict
 
 import cv2
 
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 
 
 class Node(AbstractNode):
