+++ peekingduck/runner.py
@@ -24,7 +24,7 @@ from time import perf_counter
 from typing import List
 
 from peekingduck.declarative_loader import DeclarativeLoader, NodeList
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 from peekingduck.pipeline.pipeline import Pipeline
 from peekingduck.utils.requirement_checker import RequirementChecker
