+++ tests/pipeline/test_pipeline.py
@@ -14,7 +14,7 @@
 
 import pytest
 
-from peekingduck.pipeline.nodes.node import AbstractNode
+from peekingduck.pipeline.nodes.abstract_node import AbstractNode
 from peekingduck.pipeline.pipeline import Pipeline
